Refactoring baseline commit
This commit is the *starting point* for refactoring.
It does not compile.

Changes made:
- Added Robot, Motor and Light classes
- Refactored quite a bit of control.cpp to make it much clearer
- Moved control related code into the `control` folder
- Started using enums for centering status instead of unkown sentinel values
- Changed up the frame system for printing every `n` frames
- Moved tests into their own file and separated them from the rest of the code
- Moved Encoder code to the Motor class
- Added serial_printf to more easily debug
- Started centering refactor in center_tick
- Started work on pid_tick
- And more

A lot of functionality was removed. Tests were commented out to get code to compile at some points during development
