--- include/robot/battery.h
@@ -1,6 +1,0 @@
-#ifndef CHESSBOT_BATTERY_H
-#define CHESSBOT_BATTERY_H
-
-int getBatteryLevel();
-
-#endif
