+++ include/robot/control/magnet.h
