--- include/robot/motor.h
@@ -1,8 +1,0 @@
-#ifndef CHESSBOT_MOTOR_H
-#define CHESSBOT_MOTOR_H
-
-void setupMotors();
-void setLeftPower(float power);
-void setRightPower(float power);
-
-#endif
