+++ platformio.ini
@@ -13,6 +13,9 @@ platform = espressif32
 board = lolin_s2_mini
 framework = arduino
 monitor_speed = 115200
+build_flags =
+	-Wall
+	-Wextra
 lib_deps = 
 	bblanchon/ArduinoJson@^7.2.0
 	paulstoffregen/Encoder@^1.4.4
