+++ src/tests/tests.cpp
@@ -1,0 +1,20 @@
+// #include "config.h"
+
+// void do_tests() {
+//     if (DO_DRIVE_TEST) {
+//         startDriveTest();
+//     }
+
+//     delay(2000);
+
+
+//     if (DO_DRIVE_TICKS_TEST){
+//         driveTicks(20000, "NULL");
+//     }
+        
+//     delay(5000);
+    
+//     if (DO_HARDWARE_TEST) {
+//         startMotorAndEncoderTest();
+//     }
+// }
