+++ src/utils/config.cpp
@@ -33,9 +33,11 @@ gpio_num_t PHOTODIODE_D_PIN = GPIO_NUM_6;
 
 gpio_num_t ONBOARD_LED_PIN = GPIO_NUM_15;
 gpio_num_t BATTERY_VOLTAGE_PIN = GPIO_NUM_10;
+int BATTERY_VOLTAGE_OFFSET = 100;
 
 int TICKS_PER_ROTATION = 12000;
 float TRACK_WIDTH_INCHES = 8.29;
+float TRACK_WIDTH_CM = TRACK_WIDTH_INCHES * 2.54;
 float WHEEL_DIAMETER_INCHES = 4.75;
 float THEORETICAL_MAX_VELOCITY_TPS = 52000;
 float THEORETICAL_MAX_ACCELERATION_TPSPS = 252000;
