+++ src/utils/logging.cpp
@@ -1,6 +1,8 @@
 #ifndef CHESSBOT_LOGGING_CPP
 #define CHESSBOT_LOGGING_CPP
 
+#include <stdio.h>
+
 // Associated Header File
 #include "utils/logging.h"
 
@@ -13,6 +15,29 @@
 
 // Prints a value or message through Serial. (The console)
 // ln means it sends a new newline character
+
+enum DebugLevel {
+    NONE,
+    INFO,
+    DEBUG,
+    TRACE,
+    RIDICULOUS, // Use if insane
+};
+
+void serial_printf(enum DebugLevel level, const char* fmt, ...) {
+    char buf[256];
+    
+    if level > LOGGING_LEVEL {
+        return;
+    }
+
+    va_list args;
+    va_start(args, fmt);
+
+    vsnprintf(buf, sizeof(buf) fmt, args)
+
+    Serial.print(buf)
+}
 
 void serialLog(const char *message, int serialLoggingLevel)
 {
